refactor: successHandler 및 oauth2Attribute 디렉토리 변경#55
